ensure that socket is closed upon termination of game
this was the cause of the phenomenon in which on some phones, after the
end of a game, when re-entering the SelectOpponentActivity, no players
would show. what was happening was that the onStop fcn was being invoked
late on the PokerActivity. and since the socket hadn't been closed, the
onStop fcn would close it -- but in fact it was the new socket that had
been created by the login fcn in SelectOpponentActivity.